- (*): Merged the MT changes into the main branch.
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/trunk@1662 06827809-a52a-0410-b366-d66718629ded

+++ apps/snmpdelta.c
@@ -524,7 +524,7 @@ char **argv;
 	    if (tableForm) sprintf(valueStr, "\t%.2f", printvalue);
 	    else sprintf(valueStr, " /sec: %.2f", printvalue);
 	  } else {
-	    printvalue = value;
+	    printvalue = (float)value;
 	    sprintf(valueStr, " /%d sec: %.0f",
 		    period, printvalue);
 	  }
